Changed .h extensions to .hpp Added function for conversion from bytes to memory blocks. Adddd comments. Changed mutex left out mistake in memory allocator. Abbreviated header names.

+++ src/syscall_c.c
@@ -14,7 +14,7 @@ void *mem_alloc(size_t size) {
 }
 
 int mem_free(void *addr) {
-    return (uint64) callSupervisorTrap(MEM_FREE, addr);
+    return (int) (uint64) callSupervisorTrap(MEM_FREE, addr);
 }
 
 int thread_create(thread_t *handle, void(*start_routine)(void *), void *arg) {
@@ -22,11 +22,11 @@ int thread_create(thread_t *handle, void(*start_routine)(void *), void *arg) {
     if (!stack) return -1;
 
     uint64 args[] = {(uint64) handle, (uint64) start_routine, (uint64) arg, (uint64) stack};
-    return (uint64) callSupervisorTrap(THREAD_CREATE, args);
+    return (int) (uint64) callSupervisorTrap(THREAD_CREATE, args);
 }
 
 int thread_exit() {
-    return (uint64) callSupervisorTrap(THREAD_EXIT, 0);
+    return (int) (uint64) callSupervisorTrap(THREAD_EXIT, 0);
 }
 
 void thread_dispatch() {
@@ -38,33 +38,33 @@ int thread_create_suspended(thread_t *handle, void(*start_routine)(void *), void
     if (!stack) return -1;
 
     uint64 args[] = {(uint64) handle, (uint64) start_routine, (uint64) arg, (uint64) stack};
-    return (uint64) callSupervisorTrap(THREAD_CREATE_SUSPENDED, args);
+    return (int) (uint64) callSupervisorTrap(THREAD_CREATE_SUSPENDED, args);
 }
 
 int thread_start(thread_t handle) {
-    return (uint64) callSupervisorTrap(THREAD_START, handle);
+    return (int) (uint64) callSupervisorTrap(THREAD_START, handle);
 }
 
 int sem_open(sem_t *handle, unsigned init) {
     uint64 args[] = {(uint64) handle, (uint64) init};
 
-    return (uint64) callSupervisorTrap(SEM_OPEN, args);
+    return (int) (uint64) callSupervisorTrap(SEM_OPEN, args);
 }
 
 int sem_close(sem_t handle) {
-    return (uint64) callSupervisorTrap(SEM_CLOSE, handle);
+    return (int) (uint64) callSupervisorTrap(SEM_CLOSE, handle);
 }
 
 int sem_wait(sem_t handle) {
-    return (uint64) callSupervisorTrap(SEM_WAIT, handle);
+    return (int) (uint64) callSupervisorTrap(SEM_WAIT, handle);
 }
 
 int sem_signal(sem_t handle) {
-    return (uint64) callSupervisorTrap(SEM_SIGNAL, handle);
+    return (int) (uint64) callSupervisorTrap(SEM_SIGNAL, handle);
 }
 
 int time_sleep(time_t time) {
-    return (uint64) callSupervisorTrap(TIME_SLEEP, (void *) time);
+    return (int) (uint64) callSupervisorTrap(TIME_SLEEP, (void *) time);
 }
 
 char getc() {
